requirements.txt
Flask==3.0.0
openai==1.3.7
python-dotenv==1.0.0
Pillow==10.1.0
werkzeug==3.0.1
gunicorn==21.2.0
requests==2.31.0